refactor: Add transpiler.ts (exports transpilePSVG function and PSVGFunc type)
